refactor: 뷰컨트롤러에서 coreLocation분리
